common-games-steam-friend/requirements.txt
requests_html
python-dotenv